Broke gamestate out into separate file.
